requirements.txt
fastapi
uvicorn
python-multipart
gradio
torch
torchvision
Pillow
langchain
langchain-chroma
langchain-openai
langchain-huggingface
sentence-transformers
python-dotenv
